day19.clj: make it 8x faster
The most of the speedup comes from calculating a hash
instead of having a *nested* vector `[t bots resources]`
as an entry for `seen` set.
This set can now be `dense-int-set`, which also brings improvement.

Each other improvement is relatively minor, but all together
they bring the total execution time under 600 ms.
(It was 4.7 seconds)

Big thanks to @joinr for the review, tips and general advice!